EOD cleanup

+++ crates/burn-import/src/onnx/op_configuration.rs
@@ -828,7 +828,7 @@ pub fn top_k_config(node: &Node) -> TopKConfig {
         _ => 1
     };
 
-    TopKConfig::new(axis as usize, k as usize, largest)
+    TopKConfig::new(axis as usize, k as usize, largest as usize)
 }
 
 /// Create a PadConfig from the attributes of the node
